packaging: split src into files, use /dist, use name chartjs-plugin-timeline
