+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 2-Clause License
 
-Copyright (c) 2017, Boyi C
+Copyright (c) 2021, Boyi C
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
