+++ demo/dataset-colors.html
@@ -5,14 +5,14 @@
     <title>Bar Chart</title>
     <script src="https://cdn.jsdelivr.net/npm/moment/moment.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.js"></script>
-    <script src="../timeline.js"></script>
+    <script src="../dist/chartjs-plugin-timeline.js"></script>
 
     <style>
       canvas {
+        user-select: none;
         -moz-user-select: none;
         -webkit-user-select: none;
         -ms-user-select: none;
-        user-select: none;
       }
     </style>
   </head>
