+++ package.json
@@ -1,8 +1,10 @@
 {
   "name": "chartjs-chart-timeline",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "description": "timeline chart type for chart.js",
-  "main": "src/timeline.js",
+  "main": "dist/timeline.js",
+  "jsdelivr": "dist/timeline.min.js",
+  "unpkg": "dist/timeline.min.js",
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c -w"
@@ -13,13 +15,15 @@
   },
   "peerDependencies": {
     "moment": "^2.22.0",
-    "chart.js": "^2.8.0"
+    "chart.js": "^2.9.3"
   },
   "devDependencies": {
-    "rollup": "^0.58.0",
-    "rollup-plugin-commonjs": "^9.1.0",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-uglify": "^3.0.0"
+    "rollup": "^1.0.0",
+    "rollup-plugin-commonjs": "^10.0.0",
+    "rollup-plugin-node-resolve": "^5.0.0",
+    "rollup-plugin-terser": "^5.0.0",
+    "jasmine": "^3.3.0",
+    "jasmine-core": "^3.3.0"
   },
   "keywords": [
     "chartjs",
